+++ scripts/benchmarks/benchmark_package_size.py
@@ -2,49 +2,8 @@
 
 import argparse
 import os
-import subprocess
-import httpx
 
-def get_python_version(venv_path, os_name):
-    """Get the python version of python in a virtual env.
-
-    Args:
-        venv_path: Path to virtual environment.
-        os_name: Name of os.
-
-    Returns:
-        The python version.
-    """
-    python_executable = (
-        os.path.join(venv_path, "bin", "python")
-        if "windows" not in os_name
-        else os.path.join(venv_path, "Scripts", "python.exe")
-    )
-    try:
-        output = subprocess.check_output(
-            [python_executable, "--version"], stderr=subprocess.STDOUT
-        )
-        python_version = output.decode("utf-8").strip().split()[1]
-        return ".".join(python_version.split(".")[:-1])
-    except subprocess.CalledProcessError:
-        return None
-
-
-def get_directory_size(directory):
-    """Get the size of a directory in bytes.
-
-    Args:
-        directory: The directory to check.
-
-    Returns:
-        The size of the dir in bytes.
-    """
-    total_size = 0
-    for dirpath, _, filenames in os.walk(directory):
-        for f in filenames:
-            fp = os.path.join(dirpath, f)
-            total_size += os.path.getsize(fp)
-    return total_size
+from .benchmark_utils import get_python_version, get_directory_size, send_data_to_posthog
 
 def get_package_size(venv_path, os_name):
     """Get the size of a specified package.
@@ -111,29 +70,19 @@ def insert_benchmarking_data(
         size = get_package_size(path, os_type_version)
 
     # Prepare the event data
-    event_data = {
-        "api_key": "phc_JoMo0fOyi0GQAooY3UyO9k0hebGkMyFJrrCw1Gt5SGb",
-        "event": "package_size",
-        "properties": {
-            "path": path,
-            "os": os_type_version,
-            "python_version": python_version,
-            "distinct_id": commit_sha,
-            "pr_title": pr_title,
-            "branch_name": branch_name,
-            "pr_id": pr_id,
-            "size_mb": round(size / (1024 * 1024), 3),  # save size in MB and round to 3 places
-        },
+    properties= {
+        "path": path,
+        "os": os_type_version,
+        "python_version": python_version,
+        "distinct_id": commit_sha,
+        "pr_title": pr_title,
+        "branch_name": branch_name,
+        "pr_id": pr_id,
+        "size_mb": round(size / (1024 * 1024), 3),  # save size in MB and round to 3 places
     }
 
-    # Send the data to PostHog
-    with httpx.Client() as client:
-        response = client.post("https://app.posthog.com/capture/", json=event_data)
-
-    if response.status_code != 200:
-        print(f"Error sending data to PostHog: {response.status_code} - {response.text}")
-    else:
-        print("Successfully sent benchmarking data to PostHog")
+    
+    send_data_to_posthog("package_size", properties)
 
 
 def main():
