+++ scripts/benchmarks/benchmark_web_size.py
@@ -2,24 +2,8 @@
 
 import argparse
 import os
-import httpx
 
-
-def get_directory_size(directory):
-    """Get the size of a directory in bytes.
-
-    Args:
-        directory: The directory to check.
-
-    Returns:
-        The size of the dir in bytes.
-    """
-    total_size = 0
-    for dirpath, _, filenames in os.walk(directory):
-        for f in filenames:
-            fp = os.path.join(dirpath, f)
-            total_size += os.path.getsize(fp)
-    return total_size
+from .benchmark_utils import get_directory_size, send_data_to_posthog
 
 def insert_benchmarking_data(
     os_type_version: str,
@@ -47,10 +31,7 @@ def insert_benchmarking_data(
     size = get_directory_size(path)
        
     # Prepare the event data
-    event_data = {
-        "api_key": "phc_JoMo0fOyi0GQAooY3UyO9k0hebGkMyFJrrCw1Gt5SGb",
-        "event": "web-size",
-        "properties": {
+    properties = {
             "app_name": app_name,
             "os": os_type_version,
             "python_version": python_version,
@@ -60,16 +41,8 @@ def insert_benchmarking_data(
             "pr_id": pr_id,
             "size_mb": round(size / (1024 * 1024), 3),  # save size in MB and round to 3 places
         },
-    }
 
-    # Send the data to PostHog
-    with httpx.Client() as client:
-        response = client.post("https://app.posthog.com/capture/", json=event_data)
-
-    if response.status_code != 200:
-        print(f"Error sending data to PostHog: {response.status_code} - {response.text}")
-    else:
-        print("Successfully sent benchmarking data to PostHog")
+    send_data_to_posthog("web-size", properties)
 
 
 def main():
